services/api-server/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
blinker==1.4              # via sentry-sdk
boto3==1.16.50            # via -r requirements.in
botocore==1.19.50         # via boto3, s3transfer
certifi==2020.12.5        # via requests, sentry-sdk
chardet==4.0.0            # via requests
click==7.1.2              # via flask, safety
dparse==0.5.1             # via safety
flask-cors==3.0.9         # via -r requirements.in
flask-jwt-extended==3.25.0  # via -r requirements.in
flask-script==2.0.6       # via -r requirements.in
flask==1.1.2              # via -r requirements.in, flask-cors, flask-jwt-extended, flask-script, sentry-sdk
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via requests
itsdangerous==1.1.0       # via -r requirements.in, flask
jinja2==2.11.2            # via -r requirements.in, flask
jmespath==0.10.0          # via boto3, botocore
markupsafe==1.1.1         # via jinja2
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r requirements.in
packaging==20.8           # via dparse, safety
passlib==1.7.4            # via -r requirements.in
psycopg2-binary==2.8.5    # via -r requirements.in
pyjwt==1.7.1              # via -r requirements.in, flask-jwt-extended
pyparsing==2.4.7          # via packaging
python-dateutil==2.8.1    # via botocore
python-dotenv==0.13.0     # via -r requirements.in
python-http-client==3.3.1  # via sendgrid
pyyaml==5.3.1             # via dparse
requests==2.25.1          # via safety
s3transfer==0.3.3         # via boto3
safety==1.9.0             # via -r requirements.in
sendgrid==6.4.8           # via -r requirements.in
sentry-sdk[flask]==0.19.5  # via -r requirements.in
six==1.15.0               # via flask-cors, flask-jwt-extended, python-dateutil, sqlalchemy-utils
sqlalchemy-utils==0.36.8  # via -r requirements.in
sqlalchemy==1.3.17        # via -r requirements.in, sqlalchemy-utils
starkbank-ecdsa==1.1.0    # via sendgrid
toml==0.10.2              # via dparse
typed-ast==1.4.1          # via mypy
typing-extensions==3.7.4.3  # via mypy
urllib3==1.26.2           # via botocore, requests, sentry-sdk
werkzeug==1.0.1           # via flask, flask-jwt-extended

# The following packages are considered to be unsafe in a requirements file:
# setuptools
